+++ src/engraving/libmscore/rendermidi.cpp
@@ -383,7 +383,7 @@ static void collectNote(EventMap* events, int channel, const Note* note, qreal v
 
     NoteEventList nel = note->playEvents();
     size_t nels = nel.size();
-    for (int i = 0, pitch = note->ppitch(); i < nels; ++i) {
+    for (int i = 0, pitch = note->ppitch(); i < static_cast<int>(nels); ++i) {
         const NoteEvent& e = nel[i];     // we make an explicit const ref, not a const copy.  no need to copy as we won't change the original object.
 
         // skip if note has a tie into it and only one NoteEvent
@@ -401,7 +401,7 @@ static void collectNote(EventMap* events, int channel, const Note* note, qreal v
         }
         int on  = tick1 + (ticks * e.ontime()) / 1000;
         int off = on + (ticks * e.len()) / 1000 - 1;
-        if (tieFor && i == nels - 1) {
+        if (tieFor && i == static_cast<int>(nels) - 1) {
             off += tieLen;
         }
